fixing issue where ticket status does not update when user buy ticket

+++ backend-go/module/ticket/model/ticket.go
@@ -38,7 +38,7 @@ func (TicketCreate) TableName() string { return TableName }
 type TicketUpdate struct {
 	SeatNumber int   `gorm:"column:seat_number" json:"seat_number"`
 	ShowID     int64 `gorm:"column:show_id" json:"show_id"`
-	Status     int16 `gorm:"column:status" json:"-"`
+	Status     int   `gorm:"column:status" json:"-"`
 	UserID     int64 `gorm:"column:user_id" json:"-"`
 }
